Combine passes into one
