--- lib/Conversion/JeffToTensor/CMakeLists.txt
@@ -1,15 +1,0 @@
-file(GLOB CONVERSION_SOURCES *.cpp)
-
-add_mlir_conversion_library(
-    JeffToTensor
-    ${CONVERSION_SOURCES}
-    DEPENDS
-    JeffToTensorIncGen
-    LINK_LIBS
-    MLIRArithDialect
-    MLIRDialect
-    MLIRJeff
-    MLIRTensorDialect
-    MLIRTransforms
-    DISABLE_INSTALL
-)
