+++ unittests/Conversion/test_native_round_trip.cpp
@@ -1,9 +1,5 @@
-#include "jeff/Conversion/ArithToJeff/ArithToJeff.h"
-#include "jeff/Conversion/JeffToArith/JeffToArith.h"
-#include "jeff/Conversion/JeffToMath/JeffToMath.h"
-#include "jeff/Conversion/JeffToTensor/JeffToTensor.h"
-#include "jeff/Conversion/MathToJeff/MathToJeff.h"
-#include "jeff/Conversion/TensorToJeff/TensorToJeff.h"
+#include "jeff/Conversion/JeffToNative/JeffToNative.h"
+#include "jeff/Conversion/NativeToJeff/NativeToJeff.h"
 #include "jeff/IR/JeffDialect.h"
 #include "jeff/Translation/Deserialize.hpp"
 #include "jeff/Translation/Serialize.hpp"
@@ -68,39 +64,15 @@ kj::Array<capnp::word> readJeffFile(llvm::StringRef path) {
     return capnp::messageToFlatArray(message);
 }
 
-mlir::LogicalResult convertJeffToArith(mlir::ModuleOp module) {
+mlir::LogicalResult convertJeffToNative(mlir::ModuleOp module) {
     mlir::PassManager pm(module.getContext());
-    pm.addPass(mlir::createJeffToArith());
+    pm.addPass(mlir::createJeffToNative());
     return pm.run(module);
 }
 
-mlir::LogicalResult convertJeffToMath(mlir::ModuleOp module) {
+mlir::LogicalResult convertNativeToJeff(mlir::ModuleOp module) {
     mlir::PassManager pm(module.getContext());
-    pm.addPass(mlir::createJeffToMath());
-    return pm.run(module);
-}
-
-mlir::LogicalResult convertJeffToTensor(mlir::ModuleOp module) {
-    mlir::PassManager pm(module.getContext());
-    pm.addPass(mlir::createJeffToTensor());
-    return pm.run(module);
-}
-
-mlir::LogicalResult convertArithToJeff(mlir::ModuleOp module) {
-    mlir::PassManager pm(module.getContext());
-    pm.addPass(mlir::createArithToJeff());
-    return pm.run(module);
-}
-
-mlir::LogicalResult convertMathToJeff(mlir::ModuleOp module) {
-    mlir::PassManager pm(module.getContext());
-    pm.addPass(mlir::createMathToJeff());
-    return pm.run(module);
-}
-
-mlir::LogicalResult convertTensorToJeff(mlir::ModuleOp module) {
-    mlir::PassManager pm(module.getContext());
-    pm.addPass(mlir::createTensorToJeff());
+    pm.addPass(mlir::createNativeToJeff());
     return pm.run(module);
 }
 
@@ -165,35 +137,14 @@ TEST_P(NativeRoundTripTest, RoundTrip) {
     mlirModule->print(llvm::errs());
     llvm::errs() << "\n\n";
 
-    EXPECT_TRUE(convertJeffToTensor(mlirModule.get()).succeeded());
-    EXPECT_TRUE(verify(*mlirModule).succeeded());
-
-    EXPECT_TRUE(convertJeffToArith(mlirModule.get()).succeeded());
-    EXPECT_TRUE(verify(*mlirModule).succeeded());
-
-    EXPECT_TRUE(convertJeffToMath(mlirModule.get()).succeeded());
+    EXPECT_TRUE(convertJeffToNative(mlirModule.get()).succeeded());
     EXPECT_TRUE(verify(*mlirModule).succeeded());
 
     llvm::errs() << "Converted MLIR module:\n";
     mlirModule->print(llvm::errs());
     llvm::errs() << "\n\n";
 
-    bool hasJeffOp = false;
-    mlirModule->walk([&](mlir::Operation* op) {
-        if (op->getDialect()->getNamespace() == "jeff") {
-            llvm::errs() << "Found unexpected Jeff op: " << *op << "\n";
-            hasJeffOp = true;
-        }
-    });
-    EXPECT_FALSE(hasJeffOp);
-
-    EXPECT_TRUE(convertMathToJeff(mlirModule.get()).succeeded());
-    EXPECT_TRUE(verify(*mlirModule).succeeded());
-
-    EXPECT_TRUE(convertArithToJeff(mlirModule.get()).succeeded());
-    EXPECT_TRUE(verify(*mlirModule).succeeded());
-
-    EXPECT_TRUE(convertTensorToJeff(mlirModule.get()).succeeded());
+    EXPECT_TRUE(convertNativeToJeff(mlirModule.get()).succeeded());
     EXPECT_TRUE(verify(*mlirModule).succeeded());
 
     EXPECT_TRUE(canonicalize(mlirModule.get()).succeeded());
